+++ flask_logconfig/__meta__.py
@@ -5,7 +5,7 @@ __title__ = 'Flask-LogConfig'
 __summary__ = 'Flask extension for configuring Python logging module'
 __url__ = 'https://github.com/dgilland/flask-logconfig'
 
-__version__ = '0.3.1'
+__version__ = '0.4.0'
 
 __install_requires__ = ['logconfig>=0.4.0', 'flask']
 __tests_require__ = ['tox', 'logutils']